backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
supabase==2.0.2
requests==2.31.0
flask-cors==4.0.0
werkzeug==2.3.7
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
alembic==1.12.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
openai==1.3.0
flask-socketio==5.3.6
python-socketio==5.9.0
eventlet==0.33.3
gunicorn==21.2.0